+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 Suchjs add changelog since `v1.0.0`.
 
-## [v3.0.1] - 2022-02-26
+## [v3.0.2] - 2022-02-26
 
+### Fixed
+
+- Fix the data files such as `json` and `dict` files, use a cache missing the namespace. Now it will search the file from it's own namespace also the global namespace.
+## [v3.0.1] - 2022-02-26
 ### Changed
 
 - [break change] in the template literal string you need use `//${0}` instead of `/${0}`, with two slashs as a path to reference the data type appeared in template.
@@ -12,6 +16,7 @@ Suchjs add changelog since `v1.0.0`.
 
 - Now you can reference the data type out of the template like this `/template/${0}`.
 
+- Fix the `getPathInfo` get a wrong path when the path begin with `./` in deprecated version `v3.0.0`
 ### Fixed
 
 - Fix the `cascader` and `dict` type in node is defined after `loadConf`, so you can't define a new type base on them, now fixed it.